+++ src/main/java/com/swp391/JewelrySalesSystem/entity/GemPriceList.java
@@ -31,10 +31,10 @@ public class GemPriceList extends BaseEntity implements Serializable {
   private Float carat;
 
   @Column(name = "buy_price", nullable = false)
-  private Long buyPrice;
+  private Float buyPrice;
 
   @Column(name = "sell_price", nullable = false)
-  private Long sellPrice;
+  private Float sellPrice;
 
   @Column(name = "effect_date", nullable = false)
   private Long effectDate;
